TH-66 Rename function name
